refactor: ev-zkevm-types crate restructure
