+++ crates/ev-zkevm-types/src/block/mod.rs
@@ -1,0 +1,5 @@
+pub mod io;
+pub mod verifier;
+
+pub use io::*;
+pub use verifier::BlockVerifier;
