+++ crates/ev-zkevm-types/src/hyperlane/merkle.rs
@@ -139,7 +139,8 @@ mod tests {
     use alloy_primitives::Keccak256;
     use anyhow::Result;
 
-    use crate::programs::hyperlane::tree::{MerkleTree, ZERO_HASHES};
+    use super::{MerkleTree, ZERO_HASHES};
+
     #[test]
     fn test_insert() {
         let mut tree = MerkleTree::default();
@@ -149,7 +150,6 @@ mod tests {
         let root = tree.root_with_ctx(&ZERO_HASHES.map(|s| s.to_string())).unwrap();
         let expected_root = "fa252f08612271b1aeff37a319dd0dcee621cd5d52b75b974dbac4062e56a0cc";
         assert_eq!(root, expected_root);
-        /// Hash bytes using Keccak256
         fn keccak256_hash(data: &[u8]) -> Result<String> {
             let mut hasher = Keccak256::new();
             hasher.update(data);
