+++ crates/ev-zkevm-types/src/hyperlane/message.rs
@@ -94,6 +94,7 @@ pub fn decode_token_message_body(body: &[u8]) -> Result<TokenMessageBody<'_>> {
 #[cfg(test)]
 mod tests {
     use super::*;
+
     #[test]
     fn test_decode_hyperlane_message() {
         let message = hex::decode("0300000009000004d2000000000000000000000000a7578551bae89a96c3365b93493ad2d4ebcbae9700010f2c726f757465725f617070000000000000000000000000000100000000000000000000000000000000000000006a809b36caf0d46a935ee76835065ec5a8b3cea700000000000000000000000000000000000000000000000000000000000003e8").unwrap();
