--- crates/ev-zkevm-types/src/programs/hyperlane/mod.rs
@@ -1,10 +1,0 @@
-pub mod tree;
-pub mod types;
-
-use sha3::{Digest, Keccak256};
-
-pub fn digest_keccak(bytes: &[u8]) -> [u8; 32] {
-    let mut hasher = Keccak256::new();
-    hasher.update(bytes);
-    hasher.finalize().into()
-}
