--- crates/ev-zkevm-types/src/programs/mod.rs
@@ -1,2 +1,0 @@
-pub mod block;
-pub mod hyperlane;
