+++ crates/sp1/ev-hyperlane/program/src/main.rs
@@ -21,7 +21,7 @@
 use std::str::FromStr;
 
 use alloy_primitives::{hex, Address};
-use ev_zkevm_types::programs::hyperlane::types::{HyperlaneMessageInputs, HyperlaneMessageOutputs};
+use ev_zkevm_types::hyperlane::io::{HyperlaneMessageInputs, HyperlaneMessageOutputs};
 sp1_zkvm::entrypoint!(main);
 
 pub fn main() {
