+++ crates/storage/src/hyperlane/mod.rs
@@ -23,7 +23,7 @@ mod tests {
         message::HyperlaneMessageStore,
         snapshot::{HyperlaneSnapshot, HyperlaneSnapshotStore},
     };
-    use ev_zkevm_types::{hyperlane::decode_hyperlane_message, programs::hyperlane::tree::MerkleTree};
+    use ev_zkevm_types::{hyperlane::decode_hyperlane_message, hyperlane::merkle::MerkleTree};
     use tempfile::TempDir;
 
     const DEFAULT_MESSAGE: &str = "0300000009000004d2000000000000000000000000a7578551bae89a96c3365b93493ad2d4ebcbae9700010f2c726f757465725f617070000000000000000000000000000100000000000000000000000000000000000000006a809b36caf0d46a935ee76835065ec5a8b3cea700000000000000000000000000000000000000000000000000000000000003e8";
