+++ testdata/vkeys/ev-batch-vkey-hash
@@ -1,1 +1,1 @@
-0x00192d033f9e9ee0c45596a91ce89eb346c75fb3fb2d9b615e2f5616c0be5c8a
+0x00d77e66885d5d748862ba10a691794e244e806c6579496248c8d996ca722911
