+++ testdata/vkeys/ev-hyperlane-vkey-hash
@@ -1,1 +1,1 @@
-0x00070548e48731b93d3b335eaead3ea4c734c3f0367d55246ab9484a83cf6664
+0x00c351c23825e7088f13d4a7a1585f7d271faeceba289e1c5c4e85bd514a81b7
